Make sure the invalid date in this test actually is invalid.
Previous versions of this test program used "31 dec 99999999" as an example of
an invalid date that should fail the parser, but as it happens parsing this
date works fine when 'int' -- the type used in timestamp::tm_year -- is 64 bit
large. To make sure the test doesn't unexpectedly succeed, a few additional
digits were added.