mailbox_parser: fix grammar for display-name in angle addresses
RFC2822 section 3.2.6 defines a phrase to be a (non-empty) list of words, which
is the syntax previous versions of the mailbox_parser implemented. The RFC also
specifies, however, that obsolete phrase syntax is legal too:

    obs_phrase = word *( word | "." | CFWS )

This patch adds support for that syntax to allow mailbox specifications such as 

    Dr. Foo Bar <foo.bar@example.org>

to be parsed correctly. Thanks to Stefan Baehring for reporting that problem.